+++ src/Capabilities/SettingController.h
@@ -1,16 +1,23 @@
 #pragma once
 
+#include <variant>
+
 #include "../SinricProRequest.h"
 #include "../SinricProStrings.h"
 
 #include "../SinricProNamespace.h"
 namespace SINRICPRO_NAMESPACE {
 
-using SetSettingCallback = std::function<bool(const String&, const String&, String&)>;
-
 FSTR(SETTING, setSetting);     // "setSetting"
 FSTR(SETTING, id);             // "id"
 FSTR(SETTING, value);          // "value"
+
+/**
+ * @brief Variant type for setting values that can hold int, float, bool, or String
+ */
+using SettingValue = std::variant<int, float, bool, String>;
+
+using SetSettingCallback = std::function<bool(const String&, const String&, SettingValue&)>;
 
 template <typename T>
 class SettingController {
@@ -21,14 +28,14 @@ class SettingController {
   protected:
     bool handleSettingController(SinricProRequest &request);
 
-  private: 
+  private:
     SetSettingCallback setSettingCallback;
 };
 
 template <typename T>
-SettingController<T>::SettingController() { 
+SettingController<T>::SettingController() {
   T* device = static_cast<T*>(this);
-  device->registerRequestHandler(std::bind(&SettingController<T>::handleSettingController, this, std::placeholders::_1)); 
+  device->registerRequestHandler(std::bind(&SettingController<T>::handleSettingController, this, std::placeholders::_1));
 }
 
 template <typename T>
@@ -43,11 +50,41 @@ bool SettingController<T>::handleSettingController(SinricProRequest &request) {
   bool success = false;
 
   if (setSettingCallback && request.action == FSTR_SETTING_setSetting) {
-    String settingId    = request.request_value[FSTR_SETTING_id] | "";
-    String settingValue = request.request_value[FSTR_SETTING_value] | "";
+    String settingId = request.request_value[FSTR_SETTING_id] | "";
+    JsonVariant valueVariant = request.request_value[FSTR_SETTING_value];
+
+    SettingValue settingValue;
+
+    if (valueVariant.is<bool>()) {
+      settingValue = valueVariant.as<bool>();
+    } else if (valueVariant.is<float>()) {
+      settingValue = valueVariant.as<float>();
+    } else if (valueVariant.is<int>()) {
+      settingValue = valueVariant.as<int>();
+    } else if (valueVariant.is<JsonObject>()) {
+      String jsonString;
+      serializeJson(valueVariant, jsonString);
+      settingValue = jsonString;
+    } else if (valueVariant.is<const char*>()) {
+      settingValue = String(valueVariant.as<const char*>());
+    }
+
     success = setSettingCallback(device->deviceId, settingId, settingValue);
-    request.response_value[FSTR_SETTING_id]    = settingId;
-    request.response_value[FSTR_SETTING_value] = settingValue;
+
+    request.response_value[FSTR_SETTING_id] = settingId;
+
+    if (valueVariant.is<JsonObject>()) {
+      request.response_value[FSTR_SETTING_value] = valueVariant;
+    } else if (std::holds_alternative<int>(settingValue)) {
+      request.response_value[FSTR_SETTING_value] = std::get<int>(settingValue);
+    } else if (std::holds_alternative<float>(settingValue)) {
+      request.response_value[FSTR_SETTING_value] = std::get<float>(settingValue);
+    } else if (std::holds_alternative<bool>(settingValue)) {
+      request.response_value[FSTR_SETTING_value] = std::get<bool>(settingValue);
+    } else if (std::holds_alternative<String>(settingValue)) {
+      request.response_value[FSTR_SETTING_value] = std::get<String>(settingValue);
+    }
+
     return success;
   }
 
@@ -55,6 +92,8 @@ bool SettingController<T>::handleSettingController(SinricProRequest &request) {
 }
 
 } // SINRICPRO_NAMESPACE
+
+using SettingValue = SINRICPRO_NAMESPACE::SettingValue;
 
 template <typename T>
 using SettingController = SINRICPRO_NAMESPACE::SettingController<T>;